move "get_dek_from_vault" from jwe to app. https://github.com/p15r/distributey/issues/55
